reorganized and cleaned
